requirements.txt
pytest==7.4.0
selene>=2.0.0b14
allure-pytest==2.13.2